backend/requirements.txt
fastapi==0.135.1
python-dotenv==1.0.1
uvicorn==0.35.0
pandas>=1.3.0
requests>=2.28.0
gspread>=5.4.0
google-auth>=2.6.0
google-api-python-client>=2.100.0
selenium>=4.10.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
email-validator>=2.0.0
firebase-admin>=6.2.0
pydantic==2.12.5
SQLAlchemy==2.0.48
httpx==0.28.1
